requirements.txt
django-bootstrap-toolkit
distribute>=0.6.28
django-openid-auth
pymongo
python-gcm
django-uni-form
